Rename clusterInfo
