+++ src/routes/details/ocpBreakdown/clusterInfo/clusterInfo.styles.ts
