+++ src/routes/details/ocpBreakdown/clusterInfo/clusterInfo.tsx
