+++ src/routes/details/ocpBreakdown/clusterInfo/clusterInfoContent.scss
