+++ src/routes/details/ocpBreakdown/clusterInfo/index.ts
